requirements.txt
absl-py==0.7.1
asciitree==0.3.3
astor==0.8.0
audioread==2.1.8
cffi==1.12.3
cloudpickle==1.2.1
ConfigSpace==0.4.10
cycler==0.10.0
Cython==0.29.11
dask==2.2.0
decorator==4.4.0
fasteners==0.15
fsspec==0.4.1
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.5
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
hpbandster==0.7.4
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.6.3
llvmlite==0.29.0
locket==0.2.0
Markdown==3.1.1
matplotlib==3.1.1
monotonic==1.5
netifaces==0.10.9
numba==0.44.1
numpy==1.17.0
pandas==0.25.0
partd==1.0.0
patsy==0.5.1
protobuf==3.8.0
pycparser==2.19
pyparsing==2.4.0
Pyro4==4.76
pytest-runner==5.1
python-dateutil==2.8.0
pytz==2019.1
resampy==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
serpent==1.28
six==1.12.0
smmap2==2.0.5
sounddevice==0.3.13
SoundFile==0.10.2
statsmodels==0.10.0
tb-nightly==1.14.0a20190603
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
toolz==0.9.0
tqdm==4.32.2
typing==3.7.4
Werkzeug==0.15.4
wrapt==1.11.2
